env_tic_requirements.txt
altgraph==0.17.4
anyio==4.11.0
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
arrow==1.4.0
asttokens==3.0.0
async-lru==2.0.5
attrs==25.4.0
babel==2.17.0
beautifulsoup4==4.13.5
bleach==6.3.0
blinker==1.9.0
boto3==1.38.32
botocore==1.38.32
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
comm==0.2.2
contourpy==1.3.3
cryptography==45.0.6
cycler==0.12.1
debugpy==1.8.14
decorator==5.2.1
deep-translator==1.11.4
defusedxml==0.7.1
emoji==2.14.1
et_xmlfile==2.0.0
executing==2.2.0
fastjsonschema==2.21.2
Flask==3.1.1
fonttools==4.59.0
fqdn==1.5.1
ftfy==6.3.1
gensim==4.3.3
greenlet==3.2.4
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
importlib_metadata==8.7.0
ipykernel==7.1.0
ipython==9.3.0
ipython_pygments_lexers==1.1.1
ipywidgets==8.1.8
isoduration==20.11.0
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
jmespath==1.0.1
joblib==1.5.1
json5==0.12.1
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter==1.1.1
jupyter-cache==1.0.1
jupyter-console==6.6.3
jupyter-events==0.12.0
jupyter-lsp==2.3.0
jupyter_client==8.6.3
jupyter_core==5.8.1
jupyter_server==2.17.0
jupyter_server_terminals==0.5.3
jupyterlab==4.4.10
jupyterlab_pygments==0.3.0
jupyterlab_server==2.28.0
jupyterlab_widgets==3.0.16
kiwisolver==1.4.8
langdetect==1.0.9
lark==1.3.1
MarkupSafe==3.0.2
matplotlib==3.10.3
matplotlib-inline==0.1.7
mistune==3.1.4
narwhals==2.10.2
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.5
nltk==3.9.1
notebook==7.4.7
notebook_shim==0.2.4
numpy==1.26.4
openpyxl==3.1.5
overrides==7.7.0
packaging==25.0
pandas==2.3.0
pandocfilters==1.5.1
parso==0.8.4
pdfminer.six==20250506
pdfplumber==0.11.7
pefile==2023.2.7
pillow==11.3.0
platformdirs==4.3.8
plotly==6.4.0
prometheus_client==0.23.1
prompt_toolkit==3.0.51
psutil==7.0.0
pure_eval==0.2.3
pyarrow==21.0.0
pycparser==2.22
Pygments==2.19.1
pyinstaller==6.15.0
pyinstaller-hooks-contrib==2025.8
PyMuPDF==1.26.4
pyparsing==3.2.3
pypdfium2==4.30.0
python-dateutil==2.9.0.post0
python-json-logger==4.0.0
pytz==2025.2
pywin32==310
pywin32-ctypes==0.2.3
pywinpty==3.0.2
PyYAML==6.0.2
pyzmq==26.4.0
RapidFuzz==3.14.1
referencing==0.37.0
regex==2024.11.6
requests==2.32.4
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rfc3987-syntax==1.1.0
rpds-py==0.28.0
s3transfer==0.13.0
scikit-learn==1.7.0
scipy==1.13.1
seaborn==0.13.2
Send2Trash==1.8.3
six==1.17.0
smart-open==7.1.0
sniffio==1.3.1
soupsieve==2.8
SQLAlchemy==2.0.44
stack-data==0.6.3
tabulate==0.9.0
terminado==0.18.1
threadpoolctl==3.6.0
tinycss2==1.4.0
tornado==6.5.1
tqdm==4.67.1
traitlets==5.14.3
typing_extensions==4.14.0
tzdata==2025.2
unicodedata2==16.0.0
uri-template==1.3.0
urllib3==2.4.0
wcwidth==0.2.13
webcolors==25.10.0
webencodings==0.5.1
websocket-client==1.9.0
Werkzeug==3.1.3
widgetsnbextension==4.0.15
wrapt==1.17.2
xlrd==2.0.2
xlsxwriter==3.2.5
zipp==3.23.0
